+++ imblearn/over_sampling/tests/test_adasyn.py
@@ -8,6 +8,7 @@ from numpy.testing import (assert_array_almost_equal, assert_array_equal,
                            assert_equal, assert_raises, assert_warns)
 from sklearn.datasets import make_classification
 from sklearn.utils.estimator_checks import check_estimator
+from sklearn.neighbors import NearestNeighbors
 
 from imblearn.over_sampling import ADASYN
 
@@ -217,3 +218,50 @@ def test_multiclass_error():
     y = np.array([0] * 3 + [1] * 2 + [2] * 15)
     ada = ADASYN(random_state=RND_SEED)
     assert_warns(UserWarning, ada.fit, X, y)
+
+
+def test_ada_fit_sample_nn_obj():
+    """Test fit-sample with nn object"""
+
+    # Resample the data
+    nn = NearestNeighbors(n_neighbors=6)
+    ada = ADASYN(random_state=RND_SEED, n_neighbors=nn)
+    X_resampled, y_resampled = ada.fit_sample(X, Y)
+
+    X_gt = np.array([[0.11622591, -0.0317206],
+                     [0.77481731, 0.60935141],
+                     [1.25192108, -0.22367336],
+                     [0.53366841, -0.30312976],
+                     [1.52091956, -0.49283504],
+                     [-0.28162401, -2.10400981],
+                     [0.83680821, 1.72827342],
+                     [0.3084254, 0.33299982],
+                     [0.70472253, -0.73309052],
+                     [0.28893132, -0.38761769],
+                     [1.15514042, 0.0129463],
+                     [0.88407872, 0.35454207],
+                     [1.31301027, -0.92648734],
+                     [-1.11515198, -0.93689695],
+                     [-0.18410027, -0.45194484],
+                     [0.9281014, 0.53085498],
+                     [-0.14374509, 0.27370049],
+                     [-0.41635887, -0.38299653],
+                     [0.08711622, 0.93259929],
+                     [1.70580611, -0.11219234],
+                     [0.29427267, 0.21740707],
+                     [0.68118697, -0.25220353],
+                     [1.37180201, 0.37279378],
+                     [-0.59243851, -0.80715327]])
+    y_gt = np.array([0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1,
+                     0, 0, 0, 0, 0])
+    assert_array_almost_equal(X_resampled, X_gt)
+    assert_array_equal(y_resampled, y_gt)
+
+
+def test_ada_wrong_nn_obj():
+    """Test either if an error is raised while passing a wrong NN object"""
+
+    # Resample the data
+    nn = 'rnd'
+    ada = ADASYN(random_state=RND_SEED, n_neighbors=nn)
+    assert_raises(ValueError, ada.fit_sample, X, Y)