+++ imblearn/over_sampling/tests/test_smote.py
@@ -8,6 +8,7 @@ from numpy.testing import (assert_array_almost_equal, assert_array_equal,
                            assert_equal, assert_raises, assert_warns)
 from sklearn.datasets import make_classification
 from sklearn.utils.estimator_checks import check_estimator
+from sklearn.neighbors import NearestNeighbors
 
 from imblearn.over_sampling import SMOTE
 
@@ -335,3 +336,113 @@ def test_multiclass_error():
     y = np.array([0] * 3 + [1] * 2 + [2] * 15)
     sm = SMOTE(random_state=RND_SEED)
     assert_warns(UserWarning, sm.fit, X, y)
+
+
+def test_fit_sample_nn_obj():
+    """Test sample with NN object provided."""
+
+    # Create the object
+    kind = 'borderline1'
+    nn_m = NearestNeighbors(n_neighbors=11)
+    nn_k = NearestNeighbors(n_neighbors=6)
+    smote = SMOTE(random_state=RND_SEED, kind=kind,
+                  k_neighbors=nn_k, m_neighbors=nn_m)
+
+    X_resampled, y_resampled = smote.fit_sample(X, Y)
+
+    X_gt = np.array([[0.11622591, -0.0317206],
+                     [0.77481731, 0.60935141],
+                     [1.25192108, -0.22367336],
+                     [0.53366841, -0.30312976],
+                     [1.52091956, -0.49283504],
+                     [-0.28162401, -2.10400981],
+                     [0.83680821, 1.72827342],
+                     [0.3084254, 0.33299982],
+                     [0.70472253, -0.73309052],
+                     [0.28893132, -0.38761769],
+                     [1.15514042, 0.0129463],
+                     [0.88407872, 0.35454207],
+                     [1.31301027, -0.92648734],
+                     [-1.11515198, -0.93689695],
+                     [-0.18410027, -0.45194484],
+                     [0.9281014, 0.53085498],
+                     [-0.14374509, 0.27370049],
+                     [-0.41635887, -0.38299653],
+                     [0.08711622, 0.93259929],
+                     [1.70580611, -0.11219234],
+                     [0.3765279, -0.2009615],
+                     [0.55276636, -0.10550373],
+                     [0.45413452, -0.08883319],
+                     [1.21118683, -0.22817957]])
+    y_gt = np.array([0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1,
+                     0, 0, 0, 0, 0])
+    assert_array_almost_equal(X_resampled, X_gt)
+    assert_array_equal(y_resampled, y_gt)
+
+
+def test_sample_regular_with_nn():
+    """Test sample function with regular SMOTE with a NN object."""
+
+    # Create the object
+    kind = 'regular'
+    nn_k = NearestNeighbors(n_neighbors=6)
+    smote = SMOTE(random_state=RND_SEED, kind=kind,
+                  k_neighbors=nn_k)
+
+    X_resampled, y_resampled = smote.fit_sample(X, Y)
+
+    X_gt = np.array([[0.11622591, -0.0317206],
+                     [0.77481731, 0.60935141],
+                     [1.25192108, -0.22367336],
+                     [0.53366841, -0.30312976],
+                     [1.52091956, -0.49283504],
+                     [-0.28162401, -2.10400981],
+                     [0.83680821, 1.72827342],
+                     [0.3084254, 0.33299982],
+                     [0.70472253, -0.73309052],
+                     [0.28893132, -0.38761769],
+                     [1.15514042, 0.0129463],
+                     [0.88407872, 0.35454207],
+                     [1.31301027, -0.92648734],
+                     [-1.11515198, -0.93689695],
+                     [-0.18410027, -0.45194484],
+                     [0.9281014, 0.53085498],
+                     [-0.14374509, 0.27370049],
+                     [-0.41635887, -0.38299653],
+                     [0.08711622, 0.93259929],
+                     [1.70580611, -0.11219234],
+                     [0.29307743, -0.14670439],
+                     [0.84976473, -0.15570176],
+                     [0.61319159, -0.11571668],
+                     [0.66052536, -0.28246517]])
+    y_gt = np.array([0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1,
+                     0, 0, 0, 0, 0])
+    assert_array_almost_equal(X_resampled, X_gt)
+    assert_array_equal(y_resampled, y_gt)
+
+
+def test_wrong_nn():
+    """Test either if an error is raised while passing a wrong NN object."""
+
+    # Create the object
+    kind = 'borderline1'
+    nn_m = 'rnd'
+    nn_k = NearestNeighbors(n_neighbors=6)
+    smote = SMOTE(random_state=RND_SEED, kind=kind,
+                  k_neighbors=nn_k, m_neighbors=nn_m)
+
+    assert_raises(ValueError, smote.fit_sample, X, Y)
+
+    nn_k = 'rnd'
+    nn_n = NearestNeighbors(n_neighbors=10)
+    smote = SMOTE(random_state=RND_SEED, kind=kind,
+                  k_neighbors=nn_k, m_neighbors=nn_m)
+
+    assert_raises(ValueError, smote.fit_sample, X, Y)
+
+    kind = 'regular'
+    nn_k = 'rnd'
+    smote = SMOTE(random_state=RND_SEED, kind=kind,
+                  k_neighbors=nn_k)
+
+    assert_raises(ValueError, smote.fit_sample, X, Y)